Remove two fns
Remove load_schema and get_abs_path. load_schema('foo') is supposed to load the 'foo.json' file in the 'schema' directory. It does so by looking at the absolute path of __file__. In this case, that will be the absolute path of singer/utils.py, not the absolute path of the Tap that is calling this library fn. It doesn't make much sense for these functions to exist in this common library, and they've caused confusion for Tap authors.