Initialize the singer root logger once
Only initialize the logger once, and only if the _INIT_LOGGER module private
variable is True. The reason for this is to allow client applications to
override logging with custom formatters (& more) when necessary. For instance,
to use eliot[1] to log but take advantage of the metrics and timers supplied by
singer-python.

[1] eliot: https://eliot.readthedocs.io/en/stable/